Add test for adding and deleting to do
