+++ file.txt
@@ -1,0 +1,5 @@
+$ npm i jest-environment-jsdom
+
+/**
+ * @jest-environment jsdom
+ */
