+++ package.json
@@ -4,7 +4,7 @@
   "description": "",
   "main": "index.js",
   "scripts": {
-    "test": "echo \"Error: no test specified\" && exit 1",
+    "test": "jest --coverage",
     "build": "webpack --mode development",
     "start": "webpack serve --open",
     "deploy": "gh-pages -d dist"
@@ -15,10 +15,10 @@
   "homepage": "https://prius2055.github.io/to-do-list",
   "dependencies": {
     "gh-pages": "^5.0.0",
-    "webpack": "^5.76.3",
-    "webpack-cli": "^5.0.1"
+    "jest-environment-jsdom": "^29.5.0"
   },
   "devDependencies": {
+    "@babel/plugin-transform-modules-commonjs": "^7.21.2",
     "babel-eslint": "^10.1.0",
     "css-loader": "^6.7.3",
     "eslint": "^7.32.0",
@@ -26,11 +26,14 @@
     "eslint-plugin-import": "^2.27.5",
     "hint": "^7.1.7",
     "html-webpack-plugin": "^5.5.0",
+    "jest": "^29.5.0",
     "style-loader": "^3.3.2",
     "stylelint": "^13.13.1",
     "stylelint-config-standard": "^21.0.0",
     "stylelint-csstree-validator": "^1.9.0",
     "stylelint-scss": "^3.21.0",
-    "webpack-dev-server": "^4.13.1"
+    "webpack": "^5.78.0",
+    "webpack-cli": "^5.0.1",
+    "webpack-dev-server": "^4.13.2"
   }
 }
